Add dry option (dirty way)

+++ remove-nefs.sh
@@ -6,7 +6,7 @@
 # - better argument checks
 
 # chekc if arguments was passed
-if [ $# -lt 3 ] ; then
+if [ $# -lt 4 ] ; then
   echo "Missing arguments. Provide directory to check, extension to keep and extension to delete";
   exit 1;
 fi
@@ -15,6 +15,7 @@ fi
 A_DIR=$1
 A_KEEP_EXT=$2
 A_DEL_EXT=$3
+A_DRY=$4
 
 # check if a directory was passed as argument
 if [ ! -d "${A_DIR}" ] ; then 
@@ -35,7 +36,11 @@ for f in ${CHECK_FOR} ; do
       echo "$CHECK_FILE" exists. Keeping ${DEL_FILE}
     else
       echo "$CHECK_FILE" doesn\'t exist. Deleting ${DEL_FILE}
-      rm -fv ${DEL_FILE}
+      
+			if [ ! ${A_DRY} == "dry" ]; then 
+				echo "DELETING"
+				rm -fv ${DEL_FILE}
+			fi
     fi
   fi
